put formatting functions back into the chart helper file
